Split Init command code into smaller pieces
